FTicket/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy>=1.4.0
asyncpg>=0.27.0
alembic>=1.11.0
pydantic>=1.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.5
qrcode[pil]>=7.4
python-multipart>=0.0.6
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pydantic-settings>=2.0.0
